% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_drivers.R
\name{read_drivers}
\alias{read_drivers}
\title{Scrape licensed drivers data from the Brazilian Ministry of Transport.}
\usage{
read_drivers()
}
\value{
A \code{tibble} dataframe.
}
\description{
Scrape licensed drivers data from the Brazilian Ministry of Transport.
}
\examples{
x <- read_drivers()
}
